Aggiunto Hud all'interfaccia di gioco e immagini allo sprite, rimosso classi inutilizzate Signed-off-by: Dallas <emanuele.dallara@studio.unibo.it>

+++ BomberMan/src/main/java/it/bomberman/entity/creature/Creature.java
@@ -32,7 +32,7 @@ public abstract class Creature extends Entity {
 		this.health = health;
 	}
 
-	public float getSpeed() {
+	public int getSpeed() {
 		return speed;
 	}
